Disable OpenAPI Generator (#301)
One may run it manually to generate classes from JSON Schema

## Motivation and Context

Most of the generated classes could not be used in the current API
today.

On the other hand, it is useful to be able to generate such classes to
simplify manual API synchronization. So, Gradle plugin configuration is
preserved, but detached from the build pipeline.